+++ aoe2/aoe2_game_overlay.py
@@ -6,7 +6,7 @@ from threading import Event
 from PyQt5.QtWidgets import QApplication
 from PyQt5.QtCore import Qt
 
-from common.label_display import QLabelSettings, MultiQLabelDisplay, display_multi_label_tooltip
+from common.label_display import QLabelSettings, MultiQLabelWindow
 from common.useful_tools import cut_name_length, widget_x_end, widget_y_end
 from common.rts_overlay import RTSGameOverlay
 from common.build_order_tools import get_total_on_resource
@@ -67,11 +67,12 @@ class AoE2GameOverlay(RTSGameOverlay):
 
         # build order tooltip
         layout = self.settings.layout
-        self.build_order_tooltip = MultiQLabelDisplay(
+        self.build_order_tooltip = MultiQLabelWindow(
             font_police=layout.font_police, font_size=layout.font_size, image_height=layout.build_order.image_height,
-            border_size=0, vertical_spacing=layout.build_order.tooltip_vertical_spacing,
-            color_default=layout.color_default, game_pictures_folder=self.directory_game_pictures,
-            common_pictures_folder=self.directory_common_pictures)
+            border_size=layout.build_order.tooltip_border_size,
+            vertical_spacing=layout.build_order.tooltip_vertical_spacing, color_default=layout.color_default,
+            color_background=layout.build_order.tooltip_color_background, opacity=layout.build_order.tooltip_opacity,
+            game_pictures_folder=self.directory_game_pictures, common_pictures_folder=self.directory_common_pictures)
 
         self.update_panel_elements()  # update the current panel elements
 
@@ -101,8 +102,9 @@ class AoE2GameOverlay(RTSGameOverlay):
         layout = self.settings.layout
         self.build_order_tooltip.update_settings(
             font_police=layout.font_police, font_size=layout.font_size, image_height=layout.build_order.image_height,
-            border_size=0, vertical_spacing=layout.build_order.tooltip_vertical_spacing,
-            color_default=layout.color_default)
+            border_size=layout.build_order.tooltip_border_size,
+            vertical_spacing=layout.build_order.tooltip_vertical_spacing, color_default=layout.color_default,
+            color_background=layout.build_order.tooltip_color_background, opacity=layout.build_order.tooltip_opacity)
 
         self.update_panel_elements()  # update the current panel elements
 
@@ -113,6 +115,8 @@ class AoE2GameOverlay(RTSGameOverlay):
         self.match_data_stop_flag.set()
         if self.match_data_thread_id is not None:
             self.match_data_thread_id.join()
+
+        self.build_order_tooltip.close()
 
         self.close()
 
@@ -165,6 +169,9 @@ class AoE2GameOverlay(RTSGameOverlay):
 
     def next_panel(self):
         """Select the next panel"""
+
+        # clear tooltip
+        self.build_order_tooltip.clear()
 
         # saving the upper right corner position
         if self.selected_panel == PanelID.CONFIG:
@@ -302,11 +309,13 @@ class AoE2GameOverlay(RTSGameOverlay):
         """Select the previous step of the build order"""
         if (self.selected_panel == PanelID.BUILD_ORDER) and super().build_order_previous_step():
             self.update_build_order()  # update the rendering
+            self.build_order_tooltip.clear()  # clear tooltip
 
     def build_order_next_step(self):
         """Select the next step of the build order"""
         if (self.selected_panel == PanelID.BUILD_ORDER) and super().build_order_next_step():
             self.update_build_order()  # update the rendering
+            self.build_order_tooltip.clear()  # clear tooltip
 
     def select_build_order_id(self, build_order_id: int = -1):
         """Select build order ID
@@ -769,8 +778,9 @@ class AoE2GameOverlay(RTSGameOverlay):
                     tooltip, label_x, label_y = self.build_order_resources.get_hover_tooltip(
                         0, self.mouse_x - self.x(), self.mouse_y - self.y())
                     if tooltip is not None:  # valid tooltip to display
-                        display_multi_label_tooltip(self, self.build_order_tooltip, tooltip, label_x, label_y,
-                                                    self.settings.layout.build_order.tooltip_timeout)
+                        self.build_order_tooltip.display_dictionary(
+                            tooltip, self.x() + label_x, self.y() + label_y,
+                            self.settings.layout.build_order.tooltip_timeout)
 
     def timer_match_data_call(self):
         """Function called on a timer (related to match data)"""